feat(v2.0/c): UI screens — dashboard controls, achievements, settings, store, purchase aircraft, maps
- Dashboard: play/pause auto-time loop, 1×/2×/5×/10× speed chips, +1h/+8h
  buttons, collect-profit card (shown when uncollectedProfitCoins > 0),
  Settings and Store shortcuts in TopAppBar
- AchievementsScreen: tabbed by category (3 tabs), progress bar, Claim button
- SettingsScreen: language selector, auto-time toggle, speed preference
- StoreScreen: coin-pack and welcome-pack cards, "Compra em breve" snackbar
- PurchaseAircraftScreen: ScrollableTabRow per manufacturer, LazyColumn grouped
  by aircraft family, ModalBottomSheet with full stats, Comprar button
- MapScreen: Google Maps Compose with airport markers, flight polylines, layers
  ModalBottomSheet (map type + airport filter); Canvas fallback when Maps SDK absent
- MapViewModel: inject GetMapSettingsUseCase / UpdateMapSettingsUseCase
- NavGraph: add Achievements, Settings, Store, PurchaseAircraft composable routes
- Screen.kt: add Achievements, Settings, Store, PurchaseAircraft route objects
- aircraft_models.json: add family + variantCode fields to all 25 models
- AdvanceTimeUseCase: Long→Int fix for default deltaMinutes parameter
- SeedDataUseCase: seed achievements and boosters on first launch
- Tests: CollectProfitUseCaseTest (5 cases), EvaluateProgressUseCaseTest (3 cases)

+++ app/src/main/java/com/skytycoon/app/domain/usecase/AdvanceTimeUseCase.kt
@@ -36,7 +36,7 @@ class AdvanceTimeUseCase @Inject constructor(
     private val contractRepository: ContractRepository
 ) {
     // Primary entry point: accepts arbitrary deltaMinutes for auto-time loop
-    suspend operator fun invoke(deltaMinutes: Int = GameBalanceConfig.ADVANCE_STEP_MINUTES): AdvanceTimeResult =
+    suspend operator fun invoke(deltaMinutes: Int = GameBalanceConfig.ADVANCE_STEP_MINUTES.toInt()): AdvanceTimeResult =
         withContext(Dispatchers.IO) {
             val state = gameStateRepository.get().first()
                 ?: return@withContext AdvanceTimeResult(
@@ -136,7 +136,7 @@ class AdvanceTimeUseCase @Inject constructor(
 
     // Backward-compat overload for existing callers
     suspend fun invoke(fastMode: Boolean): AdvanceTimeResult =
-        invoke(if (fastMode) GameBalanceConfig.FAST_ADVANCE_STEP_MINUTES else GameBalanceConfig.ADVANCE_STEP_MINUTES)
+        invoke(if (fastMode) GameBalanceConfig.FAST_ADVANCE_STEP_MINUTES.toInt() else GameBalanceConfig.ADVANCE_STEP_MINUTES.toInt())
 
     private fun buildDailyMissions(dayNumber: Int): List<Mission> = listOf(
         Mission(id = 0L, type = MissionType.DAILY, title = "Fretamento Diário",
